vector-service/requirements.txt
# FastAPI and web server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Vector and ML dependencies
faiss-cpu==1.12.0
sentence-transformers==2.2.2
numpy==1.24.3
torch==2.1.1
transformers==4.35.2

# Database dependencies
# SQLite3 is included in Python standard library

# Utility dependencies
python-multipart==0.0.6
python-dotenv==1.0.0

# Development dependencies (optional)
# Testing dependencies removed - use production testing instead